Removed Meteoric Launch page, Moved membership code from payments to launch app
related to #2411
