+++ contrib/env-sample
@@ -14,7 +14,6 @@ CHAVE_PAGARME_API_PRIVADA=
 
 # Control subscriptions ads and payment.
 SUBSCRIPTIONS_OPEN=False
-METEORIC_LAUNCH_OPEN=False
 PAGSEGURO_PAYMENT_PLAN=YOUR PAYMENT PLAN ON pagseguro
 
 # Amazon S3 configuration
