--- pythonpro/payments/templates/payments/meteoric_landing_page_closed.html
@@ -1,21 +1,0 @@
-{% extends 'core/base.html' %}
-{% load static %}
-{% block title %}Período de Inscrição Encerrado{% endblock %}
-{% block body %}
-    <div class="container mt-5">
-        <div class="row">
-            <div class="col mb-5">
-                <h1 class="mb-4">Período de Inscrição da Lista de Espera Encerrado!</h1>
-                <p>Infelizmente o período para inscrição se encerrou. Mas você pode se inscrever na <a
-                        href="{% url 'launch:landing_page' %}">Semana do Programador</a>. Assim você poderá ainda se
-                    inscrever quando abrirem as matrículas regulares</p>
-                <p>Se você tiver qualquer dúvida, entre em contato pelo email <a
-                        href="mailto:{{ DEFAULT_FROM_EMAIL }}?subject=Pessoas%20Pessoais">{{ DEFAULT_FROM_EMAIL }}</a>.
-                </p>
-                <p>Muito obrigado pela confiança em nosso trabalho!</p>
-                <p>Grande abraço, <br/>
-                    Luciano Ramalho e Renzo Nuccitelli</p>
-            </div>
-        </div>
-    </div>
-{% endblock body %}
